C4267の解決

+++ Server/httplib.h
@@ -2870,12 +2870,12 @@ inline bool SocketStream::is_writable() const {
 }
 
 inline ssize_t SocketStream::read(char *ptr, size_t size) {
-  if (is_readable()) { return recv(sock_, ptr, size, 0); }
+  if (is_readable()) { return recv(sock_, ptr, static_cast<int>(size), 0); }
   return -1;
 }
 
 inline ssize_t SocketStream::write(const char *ptr, size_t size) {
-  if (is_writable()) { return send(sock_, ptr, size, 0); }
+  if (is_writable()) { return send(sock_, ptr, static_cast<int>(size), 0); }
   return -1;
 }
